Add back post delete, add delete confirmation

+++ client/src/components/PostComment.tsx
@@ -135,7 +135,9 @@ const PostComment = (props: { comment: Comment }) => {
                   aria-label="Delete Comment"
                   onClick={(e) => {
                     e.preventDefault();
-                    deleteComment();
+                    if (window.confirm("Are you sure you want to delete this comment?")) {
+                      deleteComment();
+                    }
                   }}
                 />
               </>
